requirements.txt
pytest>=7.0
pytest-xdist>=2.5
pytest-asyncio>=0.21
pytest-timeout>=2.0
python-digitalocean>=1.17
paramiko>=3.0
requests>=2.28
pyyaml>=6.0
httpx>=0.25
